+++ src/components/App/App.js
@@ -5,7 +5,7 @@ import MovieDetails from "../MovieDetails/MovieDetails.js";
 import { Route } from 'react-router-dom'
 import "./App.css";
 import { scrubMovieData } from "../utilities.js";
-import { testAPI } from "../api.js"
+import { testAPI } from "../../api.js"
 
 
 class App extends Component {
